% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.R
\docType{data}
\name{gr}
\alias{gr}
\title{GRanges from maser event}
\format{
A GRanges object with event ranges and 3 metadata columns:
\describe{
\item{ranges}{IRanges}
\item{ID}{rMATS event ID}
\item{GeneID}{ENSG}
\item{geneSymbol}{gene symbol}
}
}
\usage{
gr
}
\description{
A dataset containing the genomic ranges associated with a maser event type
for a single gene.
}
\keyword{datasets}
